tracing: systemview: cleanup headers
Cleanup systemview headers and move sysview hooks into sysview_config.c
