--- subsys/tracing/sysview/SEGGER_SYSVIEW_Zephyr.h
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2018 Intel Corporation
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef SYSVIEW_ZEPHYR_H
-#define SYSVIEW_ZEPHYR_H
-
-#include <SEGGER_SYSVIEW.h>
-
-/* Services provided to SYSVIEW by Zephyr */
-extern const SEGGER_SYSVIEW_OS_API SYSVIEW_X_OS_TraceAPI;
-
-#endif
