+++ subsys/tracing/sysview/sysview.c
@@ -9,14 +9,7 @@
 #include <ksched.h>
 
 #include <SEGGER_SYSVIEW.h>
-#include <Global.h>
-#include "SEGGER_SYSVIEW_Zephyr.h"
 
-#if CONFIG_THREAD_MAX_NAME_LEN
-#define THREAD_NAME_LEN CONFIG_THREAD_MAX_NAME_LEN
-#else
-#define THREAD_NAME_LEN 20
-#endif
 
 static uint32_t interrupt;
 
@@ -72,70 +65,41 @@ void sys_trace_idle(void)
 	SEGGER_SYSVIEW_OnIdle();
 }
 
-static void set_thread_name(char *name, struct k_thread *thread)
+void sys_trace_semaphore_init(struct k_sem *sem)
 {
-	const char *tname = k_thread_name_get(thread);
-
-	if (tname != NULL && tname[0] != '\0') {
-		memcpy(name, tname, THREAD_NAME_LEN);
-		name[THREAD_NAME_LEN - 1] = '\0';
-	} else {
-		snprintk(name, THREAD_NAME_LEN, "T%pE%p",
-		thread, &thread->entry);
-	}
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_SEMA_INIT, (uint32_t)(uintptr_t)sem);
 }
 
-void sys_trace_thread_info(struct k_thread *thread)
+void sys_trace_semaphore_take(struct k_sem *sem)
 {
-	char name[THREAD_NAME_LEN];
-
-	set_thread_name(name, thread);
-
-	SEGGER_SYSVIEW_TASKINFO Info;
-
-	Info.TaskID = (uint32_t)(uintptr_t)thread;
-	Info.sName = name;
-	Info.Prio = thread->base.prio;
-	Info.StackBase = thread->stack_info.size;
-	Info.StackSize = thread->stack_info.start;
-	SEGGER_SYSVIEW_SendTaskInfo(&Info);
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_SEMA_TAKE, (uint32_t)(uintptr_t)sem);
 }
 
-static void send_task_list_cb(void)
+void sys_trace_semaphore_give(struct k_sem *sem)
 {
-	struct k_thread *thread;
-
-	for (thread = _kernel.threads; thread; thread = thread->next_thread) {
-		char name[THREAD_NAME_LEN];
-
-		if (z_is_idle_thread_object(thread)) {
-			continue;
-		}
-
-		set_thread_name(name, thread);
-
-		SEGGER_SYSVIEW_SendTaskInfo(&(SEGGER_SYSVIEW_TASKINFO) {
-			.TaskID = (uint32_t)(uintptr_t)thread,
-			.sName = name,
-			.StackSize = thread->stack_info.size,
-			.StackBase = thread->stack_info.start,
-			.Prio = thread->base.prio,
-		});
-	}
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_SEMA_GIVE, (uint32_t)(uintptr_t)sem);
 }
 
-
-static U64 get_time_cb(void)
+void sys_trace_mutex_init(struct k_mutex *mutex)
 {
-	return (U64)k_cycle_get_32();
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_MUTEX_INIT, (uint32_t)(uintptr_t)mutex);
 }
 
+void sys_trace_mutex_lock(struct k_mutex *mutex)
+{
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_MUTEX_LOCK, (uint32_t)(uintptr_t)mutex);
+}
 
-const SEGGER_SYSVIEW_OS_API SYSVIEW_X_OS_TraceAPI = {
-	get_time_cb,
-	send_task_list_cb,
-};
-
+void sys_trace_mutex_unlock(struct k_mutex *mutex)
+{
+	SEGGER_SYSVIEW_RecordU32(
+		SYS_TRACE_ID_MUTEX_UNLOCK, (uint32_t)(uintptr_t)mutex);
+}
 
 static int sysview_init(const struct device *arg)
 {
